requirements.txt
Flask
Flask-WTF
Flask-Session
pymongo
dnspython
python-dotenv
Werkzeug
gunicorn
